Moving relations migration to own task
This commit includes separating the relations migration to its own task.
This is mainly done to avoid issues with migrating data due to foreign keys or other issuse with relations.
Secondly, this commit also includes an update to the format of relations, as v9 has moved relations data to a meta object.

Fix relations mapping to check in meta objects

Run relations migration before pushing data

The relations need to be added first, otherwise the items with relations of collections are not handled correctly by the API.
This commit also adds a final context saving.

fix gitignore