Save context in between tasks and fix relations migration (#37)
* Add context functionality

This commit includes a new commandline option to pass a context file.
Moreover, it provides the basis to track which steps where already completed
through the context. There will always be a default context file that is used to seed the steps.

* Skipping steps based on completedSteps in context

Adding skip functions to skip a certain step if it is already completed in the passed context file

* Writing context after migrating schema steps

* Saving context after files migration

* Save context for roles and users

* Remove step for data as context isn't modified

* Adjust behavior of m2m fields during collection creation

Currently, m2m fields in v8 collections aren't correctly created, so that they work as expected in v9.

* Change the fieldtype to UUID if the field is pointing to directus_files_id

* Moving relations migration to own task

This commit includes separating the relations migration to its own task.
This is mainly done to avoid issues with migrating data due to foreign keys or other issuse with relations.
Secondly, this commit also includes an update to the format of relations, as v9 has moved relations data to a meta object.

Fix relations mapping to check in meta objects

Run relations migration before pushing data

The relations need to be added first, otherwise the items with relations of collections are not handled correctly by the API.
This commit also adds a final context saving.

fix gitignore

* Sort collections to add actual before junction collections

Sort collections based on their relations

This commit tries to implement some logic to sort collections based on their dependencies. Unfortunately, this is likely still error prone, especially for highly complex v8 instances with lots of relations between collections.
I think we should try to implement something more robust here.

* Run prettier on index.js

* Move context steps to a subfolder

This commits moves all generated context dumps to a dedicated folder.

* Add description for usage of context commandline option

* Separate getting of relations data from migration

This commit separatess the getting of v8 relations and saves them to a separate context file. This helps users debug potential failures during relation migration.